requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
Deprecated==1.2.14
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
django-oauth-toolkit==2.3.0
django-redis==5.3.0
djangorestframework==3.14.0
Faker==19.2.0
idna==3.4
jwcrypto==1.5.0
Markdown==3.4.3
oauthlib==3.2.2
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
validate-docbr==1.10.0
wrapt==1.15.0
